update val.py for segment
