+++ segment/detect.py
@@ -33,7 +33,7 @@ import torch
 import torch.backends.cudnn as cudnn
 
 FILE = Path(__file__).resolve()
-ROOT = FILE.parents[0]  # YOLOv5 root directory
+ROOT = FILE.parents[1]  # YOLOv5 root directory
 if str(ROOT) not in sys.path:
     sys.path.append(str(ROOT))  # add ROOT to PATH
 ROOT = Path(os.path.relpath(ROOT, Path.cwd()))  # relative
@@ -42,9 +42,10 @@ from models.experimental import attempt_load
 from utils.dataloaders import IMG_FORMATS, VID_FORMATS, LoadImages, LoadStreams
 from utils.general import (LOGGER, check_file, check_img_size, check_imshow, check_requirements, colorstr, cv2,
                            increment_path, print_args, scale_coords, strip_optimizer, xyxy2xywh)
-from utils.plots import Annotator, colors, save_one_box, plot_masks
+from utils.plots import Annotator, colors, save_one_box
+from utils.segment.plots import plot_masks
 from utils.torch_utils import select_device, time_sync
-from utils.segment import non_max_suppression_masks, scale_masks, process_mask_upsample
+from utils.segment.general import non_max_suppression_masks, scale_masks, process_mask_upsample
 
 
 @torch.no_grad()
