+++ utils/loggers/__init__.py
@@ -3,7 +3,6 @@
 Logging utils
 """
 
-from ast import Import
 import os
 import warnings
 
@@ -13,7 +12,8 @@ from torch.utils.tensorboard import SummaryWriter
 
 from utils.general import colorstr, cv2, emojis
 from utils.loggers.wandb.wandb_utils import WandbLogger
-from utils.plots import plot_images, plot_results, plot_results_with_masks, plot_images_and_masks
+from utils.plots import plot_images, plot_results
+from utils.segment.plots import plot_results_with_masks, plot_images_and_masks
 from utils.torch_utils import de_parallel
 
 LOGGERS = ('csv', 'tb', 'wandb')  # text-file, TensorBoard, Weights & Biases
@@ -158,7 +158,7 @@ class Loggers():
             if ((epoch + 1) % self.opt.save_period == 0 and not final_epoch) and self.opt.save_period != -1:
                 self.wandb.log_model(last.parent, self.opt, epoch, fi, best_model=best_fitness == fi)
 
-    def on_train_end(self, last, best, plots, epoch, results, masks=False):
+    def on_train_end(self, last, best, plots, epoch, results):
         # Callback runs on training end
         # plot_results = plot_results_with_masks if masks else plot_results
         if plots:
@@ -188,174 +188,10 @@ class Loggers():
         if self.wandb:
             self.wandb.wandb_run.config.update(params, allow_val_change=True)
 
-from threading import Thread
 
-class NewLoggers:
-    """Loggers without wandb, cause I don't really use `wandb` and `wandb` related codes are noisy."""
-    def __init__(
-        self,
-        save_dir=None,
-        opt=None,
-        logger=None,
-        include=LOGGERS,
-    ):
-        self.save_dir = save_dir
-        self.opt = opt
-        self.logger = logger  # for printing results to console
-        self.include = include
-        self.keys = [
-            "train/box_loss",
-            "train/obj_loss",
-            "train/cls_loss",  # train loss
-            "metrics/precision",
-            "metrics/recall",
-            "metrics/mAP_0.5",
-            "metrics/mAP_0.5:0.95",  # metrics
-            "val/box_loss",
-            "val/obj_loss",
-            "val/cls_loss",  # val loss
-            "x/lr0",
-            "x/lr1",
-            "x/lr2",
-        ]  # params
-        self.best_keys = [
-            "best/epoch",
-            "best/precision",
-            "best/recall",
-            "best/mAP_0.5",
-            "best/mAP_0.5:0.95",
-        ]
-        for k in LOGGERS:
-            setattr(self, k, None)  # init empty logger dictionary
-        self.csv = True  # always log to csv
-
-        # TensorBoard
-        s = self.save_dir
-        if "tb" in self.include and s.exists():
-            prefix = colorstr("TensorBoard: ")
-            self.logger.info(
-                f"{prefix}Start with 'tensorboard --logdir {s.parent}', view at http://localhost:6006/"
-            )
-            self.tb = SummaryWriter(str(s))
-        try:
-            import wandb
-            from wandb import __version__
-            wandb.init(project=opt.project, name=opt.name, config=opt)
-        except ImportError:
-            wandb = None
-            pass
-        self.wandb = wandb
-        
-    def on_pretrain_routine_end(self):
-        pass
-
-    def on_train_batch_end(
-        self, ni, model, imgs, targets, masks, paths, plots, sync_bn, plot_idx
-    ):
-        # Callback runs on train batch end
-        if plots and self.save_dir.exists():
-            if ni == 0:
-                if (
-                    not sync_bn
-                ):  # tb.add_graph() --sync known issue https://github.com/ultralytics/yolov5/issues/3754
-                    with warnings.catch_warnings():
-                        warnings.simplefilter("ignore")  # suppress jit trace warning
-                        self.tb.add_graph(
-                            torch.jit.trace(
-                                de_parallel(model), imgs[0:1], strict=False
-                            ),
-                            [],
-                        )
-            if plot_idx is not None and ni in plot_idx:
-                f = self.save_dir / f"train_batch{ni}.jpg"  # filename
-                Thread(
-                    target=plot_images, args=(imgs, targets, paths, f), daemon=True
-                ).start()
-            # if ni < 3:
-            #     f = self.save_dir / f'train_batch{ni}.jpg'  # filename
-            #     Thread(target=plot_images, args=(imgs, targets, paths, f), daemon=True).start()
-
-    def on_train_epoch_end(self, epoch):
-        # Callback runs on train epoch end
-        pass
-
-    def on_val_image_end(self, imgs, targets, masks, paths):
-        # Callback runs on val image end
-        pass
-
-    def on_val_end(self):
-        # Callback runs on val end
-        pass
-
-    def on_fit_epoch_end(self, vals, epoch):
-        # Callback runs at the end of each fit (train+val) epoch
-        x = {k: v for k, v in zip(self.keys, vals)}  # dict
-        if self.csv and self.save_dir.exists():
-            file = self.save_dir / "results.csv"
-            n = len(x) + 1  # number of cols
-            s = (
-                ""
-                if file.exists()
-                else (("%20s," * n % tuple(["epoch"] + self.keys)).rstrip(",") + "\n")
-            )  # add header
-            with open(file, "a") as f:
-                f.write(s + ("%20.5g," * n % tuple([epoch] + vals)).rstrip(",") + "\n")
-
-        if self.tb:
-            for k, v in x.items():
-                self.tb.add_scalar(k, v, epoch)
-        if self.wandb:
-            wandb.log(x)
-
-    def on_model_save(self, last, epoch, final_epoch, best_fitness, fi):
-        # Callback runs on model save event
-        pass
-
-    def on_train_end(self, plots, epoch, masks=False):
-        plts = plot_results_with_masks if masks else plot_results
-        # Callback runs on training end
-        if plots and self.save_dir.exists():
-            plts(file=self.save_dir / "results.csv")  # save results.png
-        files = [
-            "results.png",
-            "confusion_matrix.png",
-            *(f"{x}_curve.png" for x in ("F1", "PR", "P", "R")),
-        ]
-        files = [
-            (self.save_dir / f) for f in files if (self.save_dir / f).exists()
-        ]  # filter
-
-        if self.tb:
-            import cv2
-
-            for f in files:
-                self.tb.add_image(
-                    f.stem, cv2.imread(str(f))[..., ::-1], epoch, dataformats="HWC"
-                )
-        if self.wandb:
-            best = self.save_dir/ "weights" / "best.pt"
-            last = self.save_dir / "weights" / "last.pt"
-            wandb.log_artifact(str(best if best.exists() else last),
-                                   type='model',
-                                   name=f'run_{self.wandb.run.id}_model',
-                                   aliases=['latest', 'best', 'stripped'])
-            self.wandb.finish()
-
-
-    def on_params_update(self):
-        # Update hyperparams or configs of the experiment
-        # params: A dict containing {param: value} pairs
-        pass
-    
-class NewLoggersMask(NewLoggers):
-    def __init__(
-        self,
-        save_dir=None,
-        opt=None,
-        logger=None,
-        include=LOGGERS,
-    ):
-        super().__init__(save_dir, opt, logger, include)
+class LoggersMask(Loggers):
+    def __init__(self, save_dir=None, weights=None, opt=None, hyp=None, logger=None, include=LOGGERS):
+        super().__init__(save_dir, weights, opt, hyp, logger, include)
         self.keys = [
             "train/box_loss",
             "train/seg_loss",  # train loss
@@ -379,50 +215,73 @@ class NewLoggersMask(NewLoggers):
         ]  # params
         self.best_keys = [
             "best/epoch",
-            "best/precision",
-            "best/recall",
-            "best/mAP_0.5",
-            "best/mAP_0.5:0.95",
+            "best/precision(B)",
+            "best/recall(B)",
+            "best/mAP_0.5(B)",
+            "best/mAP_0.5:0.95(B)",
+            "best/precision(M)",
+            "best/recall(M)",
+            "best/mAP_0.5(M)",
+            "best/mAP_0.5:0.95(M)",
         ]
 
         
-    def on_train_batch_end(self, ni, model, imgs, targets, masks, paths, plots, sync_bn):
-        # Callback runs on train batch end
-        if plots and self.save_dir.exists():
+    def on_train_batch_end(self, ni, model, imgs, targets, masks, paths, plots):
+        if plots:
             if ni == 0:
-                if (
-                    not sync_bn
-                ):  # tb.add_graph() --sync known issue https://github.com/ultralytics/yolov5/issues/3754
+                if self.tb and not self.opt.sync_bn:  # --sync known issue https://github.com/ultralytics/yolov5/issues/3754
                     with warnings.catch_warnings():
-                        warnings.simplefilter("ignore")  # suppress jit trace warning
-                        self.tb.add_graph(
-                            torch.jit.trace(
-                                de_parallel(model), imgs[0:1], strict=False
-                            ),
-                            [],
-                        )
+                        warnings.simplefilter('ignore')  # suppress jit trace warning
+                        self.tb.add_graph(torch.jit.trace(de_parallel(model), imgs[0:1], strict=False), [])
             if ni < 3:
                 f = self.save_dir / f"train_batch{ni}.jpg"  # filename
                 plot_images_and_masks(imgs, targets, masks, paths, f)
-                
 
+            if self.wandb and ni == 10:
+                files = sorted(self.save_dir.glob('train*.jpg'))
+                self.wandb.log({'Mosaics': [wandb.Image(str(f), caption=f.name) for f in files if f.exists()]})
 
-    def on_fit_epoch_end(self, vals, epoch):
+    def on_fit_epoch_end(self, vals, epoch, best_fitness, fi):
         # Callback runs at the end of each fit (train+val) epoch
-        x = {k: v for k, v in zip(self.keys, vals)}  # dict
-        if self.csv and self.save_dir.exists():
-            file = self.save_dir / "results.csv"
+        x = dict(zip(self.keys, vals))
+        if self.csv:
+            file = self.save_dir / 'results.csv'
             n = len(x) + 1  # number of cols
-            s = (
-                ""
-                if file.exists()
-                else (("%20s," * n % tuple(["epoch"] + self.keys)).rstrip(",") + "\n")
-            )  # add header
-            with open(file, "a") as f:
-                f.write(s + ("%20.5g," * n % tuple([epoch] + vals)).rstrip(",") + "\n")
+            s = '' if file.exists() else (('%20s,' * n % tuple(['epoch'] + self.keys)).rstrip(',') + '\n')  # add header
+            with open(file, 'a') as f:
+                f.write(s + ('%20.5g,' * n % tuple([epoch] + vals)).rstrip(',') + '\n')
 
         if self.tb:
             for k, v in x.items():
                 self.tb.add_scalar(k, v, epoch)
+
         if self.wandb:
-            wandb.log(x, step=epoch, commit=True)
+            if best_fitness == fi:
+                best_results = [epoch] + vals[4:12]
+                for i, name in enumerate(self.best_keys):
+                    self.wandb.wandb_run.summary[name] = best_results[i]  # log best results in the summary
+            self.wandb.log(x)
+            self.wandb.end_epoch(best_result=best_fitness == fi)
+
+    def on_train_end(self, last, best, plots, epoch, results):
+        # Callback runs on training end
+        if plots:
+            plot_results_with_masks(file=self.save_dir / 'results.csv')  # save results.png
+        files = ['results.png', 'confusion_matrix.png', *(f'{x}_curve.png' for x in ('F1', 'PR', 'P', 'R'))]
+        files = [(self.save_dir / f) for f in files if (self.save_dir / f).exists()]  # filter
+        self.logger.info(f"Results saved to {colorstr('bold', self.save_dir)}")
+
+        if self.tb:
+            for f in files:
+                self.tb.add_image(f.stem, cv2.imread(str(f))[..., ::-1], epoch, dataformats='HWC')
+
+        if self.wandb:
+            self.wandb.log(dict(zip(self.keys[4:16], results)))
+            self.wandb.log({"Results": [wandb.Image(str(f), caption=f.name) for f in files]})
+            # Calling wandb.log. TODO: Refactor this into WandbLogger.log_model
+            if not self.opt.evolve:
+                wandb.log_artifact(str(best if best.exists() else last),
+                                   type='model',
+                                   name=f'run_{self.wandb.wandb_run.id}_model',
+                                   aliases=['latest', 'best', 'stripped'])
+            self.wandb.finish_run()
