+++ utils/plots.py
@@ -644,62 +644,6 @@ class Annotator:
         # Return annotated image as array
         return np.asarray(self.im)
 
-class Visualizer(object):
-    """Visualization of one model."""
-    def __init__(self, names) -> None:
-        super().__init__()
-        self.names = names
-
-    def draw_one_img(self, img, output, vis_conf=0.4):
-        """Visualize one images.
-        
-        Args:
-            imgs (numpy.ndarray): one image.
-            outputs (torch.Tensor): one output, (num_boxes, classes+5)
-            vis_confs (float, optional): Visualize threshold.
-        Return:
-            img (numpy.ndarray): Image after visualization.           
-        """
-        if isinstance(output, list):
-            output = output[0]
-        if output is None or len(output) == 0:
-            return img
-        for (*xyxy, conf, cls) in reversed(output[:, :6]):
-            if conf < vis_conf:
-                continue
-            label = '%s %.2f' % (self.names[int(cls)], conf)
-            color = colors(int(cls))
-            plot_one_box(xyxy, img, label=label,
-                         color=color, 
-                         line_thickness=2)
-        return img
-
-    def draw_multi_img(self, imgs, outputs, vis_confs=0.4):
-        """Visualize multi images.
-        
-        Args:
-            imgs (List[numpy.array]): multi images.
-            outputs (List[torch.Tensor]): multi outputs, List[num_boxes, classes+5].
-            vis_confs (float | tuple[float], optional): Visualize threshold.
-        Return:
-            imgs (List[numpy.ndarray]): Images after visualization.           
-        """
-        if isinstance(vis_confs, float):
-            vis_confs = list(repeat(vis_confs, len(imgs)))
-        assert len(imgs) == len(outputs) == len(vis_confs)
-        for i, output in enumerate(outputs):  # detections per image
-            self.draw_one_img(imgs[i], output, vis_confs[i])
-        return imgs
-
-    def draw_imgs(self, imgs, outputs, vis_confs=0.4):
-        if isinstance(imgs, np.ndarray):
-            return self.draw_one_img(imgs, outputs, vis_confs)
-        else:
-            return self.draw_multi_img(imgs, outputs, vis_confs)
-
-    def __call__(self, imgs, outputs, vis_confs=0.4):
-        return self.draw_imgs(imgs, outputs, vis_confs)
-
 
 def hist2d(x, y, n=100):
     # 2d histogram used in labels.png and evolve.png
@@ -1078,52 +1022,6 @@ def plot_results(file="path/to/results.csv", dir="", best=True):
     plt.close()
 
 
-def plot_results_with_masks(file="path/to/results.csv", dir="", best=True):
-    # Plot training results.csv. Usage: from utils.plots import *; plot_results('path/to/results.csv')
-    save_dir = Path(file).parent if file else Path(dir)
-    fig, ax = plt.subplots(2, 8, figsize=(18, 6), tight_layout=True)
-    ax = ax.ravel()
-    files = list(save_dir.glob("results*.csv"))
-    assert len(files), f"No results.csv files found in {save_dir.resolve()}, nothing to plot."
-    for _, f in enumerate(files):
-        try:
-            data = pd.read_csv(f)
-            index = np.argmax(
-                0.9 * data.values[:, 8]
-                + 0.1 * data.values[:, 7]
-                + 0.9 * data.values[:, 12]
-                + 0.1 * data.values[:, 11],
-            )
-            s = [x.strip() for x in data.columns]
-            x = data.values[:, 0]
-            for i, j in enumerate([1, 2, 3, 4, 5, 6, 9, 10, 13, 14, 15, 16, 7, 8, 11, 12]):
-                y = data.values[:, j]
-                # y[y == 0] = np.nan  # don't show zero values
-                ax[i].plot(x, y, marker=".", label=f.stem, linewidth=2, markersize=2)
-                if best:
-                    # best
-                    ax[i].scatter(
-                        index,
-                        y[index],
-                        color="r",
-                        label=f"best:{index}",
-                        marker="*",
-                        linewidth=3,
-                    )
-                    ax[i].set_title(s[j] + f"\n{round(y[index], 5)}")
-                else:
-                    # last
-                    ax[i].scatter(x[-1], y[-1], color="r", label="last", marker="*", linewidth=3)
-                    ax[i].set_title(s[j] + f"\n{round(y[-1], 5)}")
-                # if j in [8, 9, 10]:  # share train and val loss y axes
-                #     ax[i].get_shared_y_axes().join(ax[i], ax[i - 5])
-        except Exception as e:
-            print(f"Warning: Plotting error for {f}: {e}")
-    ax[1].legend()
-    fig.savefig(save_dir / "results.png", dpi=200)
-    plt.close()
-
-
 def plot_one_box(x, img, color=None, label=None, line_thickness=None):
     import random
 
@@ -1178,218 +1076,3 @@ def feature_visualization(x, module_type, stage, n=32, save_dir=Path("runs/detec
             print(f"Saving {save_dir / f}... ({n}/{channels})")
             plt.savefig(save_dir / f, dpi=300, bbox_inches="tight")
             plt.close()
-
-
-def plot_images_and_masks(
-    images,
-    targets,
-    masks,
-    paths=None,
-    fname="images.jpg",
-    names=None,
-    max_size=640,
-    max_subplots=16,
-):
-    # Plot image grid with labels
-    # print("targets:", targets.shape)
-    # print("masks:", masks.shape)
-    # print('--------------------------')
-
-    if isinstance(images, torch.Tensor):
-        images = images.cpu().float().numpy()
-    if isinstance(targets, torch.Tensor):
-        targets = targets.cpu().numpy()
-    if isinstance(masks, torch.Tensor):
-        masks = masks.cpu().numpy()
-        masks = masks.astype(int)
-
-    # un-normalise
-    if np.max(images[0]) <= 1:
-        images *= 255
-
-    tl = 3  # line thickness
-    tf = max(tl - 1, 1)  # font thickness
-    bs, _, h, w = images.shape  # batch size, _, height, width
-    bs = min(bs, max_subplots)  # limit plot images
-    ns = np.ceil(bs ** 0.5)  # number of subplots (square)
-
-    # Check if we should resize
-    scale_factor = max_size / max(h, w)
-    if scale_factor < 1:
-        h = math.ceil(scale_factor * h)
-        w = math.ceil(scale_factor * w)
-
-    mosaic = np.full((int(ns * h), int(ns * w), 3), 255, dtype=np.uint8)  # init
-    for i, img in enumerate(images):
-        if i == max_subplots:  # if last batch has fewer images than we expect
-            break
-
-        block_x = int(w * (i // ns))
-        block_y = int(h * (i % ns))
-
-        img = img.transpose(1, 2, 0)
-        if scale_factor < 1:
-            img = cv2.resize(img, (w, h))
-
-        mosaic[block_y : block_y + h, block_x : block_x + w, :] = img
-        if len(targets) > 0:
-            idx = (targets[:, 0]).astype(int)
-            image_targets = targets[idx == i]
-
-            if masks.max() > 1.0:  # mean that masks are overlap
-                image_masks = masks[[i]]  # (1, 640, 640)
-                # convert masks (1, 640, 640) -> (n, 640, 640)
-                nl = len(image_targets)
-                index = np.arange(nl).reshape(nl, 1, 1) + 1
-                image_masks = np.repeat(image_masks, nl, axis=0)
-                image_masks = np.where(image_masks == index, 1.0, 0.0)
-            else:
-                image_masks = masks[idx == i]
-
-            boxes = xywh2xyxy(image_targets[:, 2:6]).T
-            classes = image_targets[:, 1].astype("int")
-            labels = image_targets.shape[1] == 6  # labels if no conf column
-            conf = (
-                None if labels else image_targets[:, 6]
-            )  # check for confidence presence (label vs pred)
-
-            if boxes.shape[1]:
-                if boxes.max() <= 1.01:  # if normalized with tolerance 0.01
-                    boxes[[0, 2]] *= w  # scale to pixels
-                    boxes[[1, 3]] *= h
-                elif scale_factor < 1:  # absolute coords need scale if image scales
-                    boxes *= scale_factor
-            boxes[[0, 2]] += block_x
-            boxes[[1, 3]] += block_y
-            for j, box in enumerate(boxes.T):
-                cls = int(classes[j])
-                color = colors(cls)
-                cls = names[cls] if names else cls
-                mask = image_masks[j].astype(np.bool)
-                # print(mask.shape)
-                # print(mosaic.shape)
-                if labels or conf[j] > 0.25:  # 0.25 conf thresh
-                    label = "%s" % cls if labels else "%s %.1f" % (cls, conf[j])
-                    plot_one_box(box, mosaic, label=label, color=color, line_thickness=tl)
-                    mosaic[block_y : block_y + h, block_x : block_x + w, :][mask] = mosaic[
-                        block_y : block_y + h, block_x : block_x + w, :
-                    ][mask] * 0.35 + (np.array(color) * 0.65)
-
-        # Draw image filename labels
-        if paths:
-            label = Path(paths[i]).name[:40]  # trim to 40 char
-            t_size = cv2.getTextSize(label, 0, fontScale=tl / 3, thickness=tf)[0]
-            cv2.putText(
-                mosaic,
-                label,
-                (block_x + 5, block_y + t_size[1] + 5),
-                0,
-                tl / 3,
-                [220, 220, 220],
-                thickness=tf,
-                lineType=cv2.LINE_AA,
-            )
-
-        # Image border
-        cv2.rectangle(
-            mosaic,
-            (block_x, block_y),
-            (block_x + w, block_y + h),
-            (255, 255, 255),
-            thickness=3,
-        )
-
-    if fname:
-        r = min(1280.0 / max(h, w) / ns, 1.0)  # ratio to limit image size
-        mosaic = cv2.resize(
-            mosaic, (int(ns * w * r), int(ns * h * r)), interpolation=cv2.INTER_AREA
-        )
-        # cv2.imwrite(fname, cv2.cvtColor(mosaic, cv2.COLOR_BGR2RGB))  # cv2 save
-        with Image.fromarray(mosaic) as im:
-            im.save(fname)
-    return mosaic
-
-
-def plot_images_boxes_and_masks(
-    images,
-    targets,
-    masks=None,
-    paths=None,
-    fname="images.jpg",
-    names=None,
-    max_size=640,
-    max_subplots=16,
-):
-    if masks is not None:
-        return plot_images_and_masks(images, targets, masks, paths, fname, names, max_size, max_subplots)
-    else:
-        return plot_images(images, targets, paths, fname, names, max_size, max_subplots)
-
-
-def plot_masks(img, masks, colors, alpha=0.5):
-    """
-    Args:
-        img (tensor): img on cuda, shape: [3, h, w], range: [0, 1]
-        masks (tensor): predicted masks on cuda, shape: [n, h, w]
-        colors (List[List[Int]]): colors for predicted masks, [[r, g, b] * n]
-    Return:
-        img after draw masks, shape: [h, w, 3]
-
-    transform colors and send img_gpu to cpu for the most time.
-    """
-    img_gpu = img.clone()
-    num_masks = len(masks)
-    # [n, 1, 1, 3]
-    # faster this way to transform colors
-    colors = torch.tensor(colors, device=img.device).float() / 255.0
-    colors = colors[:, None, None, :]
-    # [n, h, w, 1]
-    masks = masks[:, :, :, None]
-    masks_color = masks.repeat(1, 1, 1, 3) * colors * alpha
-    inv_alph_masks = masks * (-alpha) + 1
-    masks_color_summand = masks_color[0]
-    if num_masks > 1:
-        inv_alph_cumul = inv_alph_masks[: (num_masks - 1)].cumprod(dim=0)
-        masks_color_cumul = masks_color[1:] * inv_alph_cumul
-        masks_color_summand += masks_color_cumul.sum(dim=0)
-
-    # print(inv_alph_masks.prod(dim=0).shape) # [h, w, 1]
-    img_gpu = img_gpu.flip(dims=[0])  # filp channel for opencv
-    img_gpu = img_gpu.permute(1, 2, 0).contiguous()
-    # [h, w, 3]
-    img_gpu = img_gpu * inv_alph_masks.prod(dim=0) + masks_color_summand
-    return (img_gpu * 255).byte().cpu().numpy()
-
-def visualize(self, images, outputs, out_masks, vis_confs=0.4):
-    """Image visualize
-    if images is a List of ndarray, then will return a List.
-    if images is a ndarray, then return ndarray.
-    Args:
-        outputs: bbox+conf+cls, List[torch.Tensor(num_boxes, 6)]xB.
-        masks: binary masks, List[torch.Tensor(num_boxes, img_h, img_w)]xB.
-    """
-    ori_type = type(images)
-    # get original shape, cause self.ori_hw will be cleared
-    images = images if isinstance(images, list) else [images]
-    ori_hw = [img.shape[:2] for img in images]
-    # init the list to keep image with masks.
-    # TODO: fix this bug when output is empty.
-    masks_images = []
-    # draw masks
-    for i, output in enumerate(outputs):
-        if output is None or len(output) == 0:
-            continue
-        idx = output[:, 4] > vis_confs
-        masks = out_masks[i][idx]
-        mcolors = [colors(int(cls)) for cls in output[:, 5]]
-        # NOTE: this way to draw masks is faster,
-        # from https://github.com/dbolya/yolact
-        # image with masks, (img_h, img_w, 3)
-        img_masks = plot_masks(self.imgs[i], masks, mcolors)
-        # scale image to original hw
-        from utils.segment import scale_masks
-        img_masks = scale_masks(self.imgs[i].shape[1:], img_masks, ori_hw[i])
-        masks_images.append(img_masks)
-    # TODO: make this(ori_type stuff) clean
-    images = masks_images[0] if (len(masks_images) == 1) and type(masks_images) != ori_type else images[0]
-    return self.vis(images, outputs, vis_confs)
