+++ utils/segment/metrics.py
@@ -8,13 +8,10 @@ from easydict import EasyDict as edict
 from ..metrics import ap_per_class
 
 
-def fitness(x, masks=False):
+def fitness(x):
     # Model fitness as a weighted combination of metrics
-    if masks:
-        w = [0.0, 0.0, 0.1, 0.9, 0.0, 0.0, 0.1, 0.9]
-        return (x[:, :8] * w).sum(1)
-    w = [0.0, 0.0, 0.1, 0.9]  # weights for [P, R, mAP@0.5, mAP@0.5:0.95]
-    return (x[:, :4] * w).sum(1)
+    w = [0.0, 0.0, 0.1, 0.9, 0.0, 0.0, 0.1, 0.9]
+    return (x[:, :8] * w).sum(1)
 
 
 def ap_per_class_box_and_mask(tp_m, tp_b, conf, pred_cls, target_cls, plot=False, save_dir=".", names=(), ):
@@ -30,9 +27,9 @@ def ap_per_class_box_and_mask(tp_m, tp_b, conf, pred_cls, target_cls, plot=False
             prefix="Mask")[2:]
 
     results = edict({
-        "boxes": {"p": results_boxes[0], "r": results_boxes[1], "ap": results_boxes[2], "f1": results_boxes[3],
+        "boxes": {"p": results_boxes[0], "r": results_boxes[1], "ap": results_boxes[3], "f1": results_boxes[2],
             "ap_class": results_boxes[4]},
-        "masks": {"p": results_masks[0], "r": results_masks[1], "ap": results_masks[2], "f1": results_masks[3],
+        "masks": {"p": results_masks[0], "r": results_masks[1], "ap": results_masks[3], "f1": results_masks[2],
             "ap_class": results_masks[4]}})
     return results
